#231 Refactor: Rename folder types to interfaces
