+++ typescript/big-practice/src/scripts/controllers/noteController.ts
@@ -1,5 +1,5 @@
 import NoteModel from '../models/noteModel';
-import Note from '../types/note';
+import Note from '../interfaces/note';
 import { renderPopupError } from '../utils/errorsDOM';
 import LoadingPage from '../utils/loadingPage';
 import ListNoteView from '../views/listNoteView';
