+++ typescript/big-practice/src/scripts/interfaces/note.ts
