+++ typescript/big-practice/src/scripts/interfaces/user.ts
